Restructure dir.
