+++ packages/editor/src/client/blocks/richTextBlock/tableMenu.less
@@ -2,7 +2,7 @@ table {
   white-space: pre-wrap;
   width: 100%;
   margin: 10px;
-  /* cannot defined in tsx file */
+  /* cannot be defined in tsx file */
   table-layout: fixed;
 }
